feat: conecta painel vue ao contrato da api
