+++ src/frontend/src/components/RoadmapPanel.vue
@@ -1,0 +1,18 @@
+<template>
+  <aside class="panel">
+    <div class="panel-header">
+      <div>
+        <span class="eyebrow">Roadmap</span>
+        <h2>Proximas entregas</h2>
+      </div>
+      <p>Evolucao sugerida logo apos a estrutura inicial.</p>
+    </div>
+
+    <ol class="roadmap-list">
+      <li>Persistir ativos e oportunidades em MySQL.</li>
+      <li>Criar ingestion manual e agendada de sinais.</li>
+      <li>Adicionar filtros por setor, score e upside.</li>
+      <li>Incluir autenticacao e watchlists por usuario.</li>
+    </ol>
+  </aside>
+</template>
