+++ src/frontend/src/main.js
@@ -1,5 +1,5 @@
 import { createApp } from "vue";
 import App from "./App.vue";
-import "./style.css";
+import "./styles/main.css";
 
 createApp(App).mount("#app");
