+++ src/frontend/src/router/index.js
@@ -1,0 +1,13 @@
+import DashboardPage from "../pages/DashboardPage.vue";
+
+export const routes = Object.freeze([
+  {
+    name: "dashboard",
+    path: "/",
+    component: DashboardPage,
+  },
+]);
+
+export function resolveCurrentRoute() {
+  return routes[0];
+}
